requirements.txt
langchain>=0.3.8
langchain-core>=0.3.12
langchain-community>=0.3.8
langchain-openai>=0.3.1
langchain-anthropic>=0.1.27
langchain-cohere>=0.1.10
langchain-mistralai>=0.1.16
langchain-groq>=0.1.8
pydantic>=2.7.0
python-dotenv>=1.0.1
